requirements.txt
av==14.3.0
dill==0.4.0
multiprocess==0.70.15
numpy==1.26.4
opencv-python==4.9.0.80
pillow==11.2.1
python-dotenv==1.0.0
redis==5.0.1
setproctitle==1.3.6
shortuuid==1.0.11